delete useLoggableEventsForUser
