+++ src/components/LoggableEventsGQL.tsx
@@ -1,12 +1,33 @@
+import { gql, useQuery } from '@apollo/client';
 import invariant from 'tiny-invariant';
 
+import LoggableEventCard from './EventCards/LoggableEventCard';
+import EventLabel from './EventLabels/EventLabel';
 import LoggableEventsView from './LoggableEventsView';
 
-import { useLoggableEventsForUser } from '../hooks/useLoggableEventsForUser';
 import { useAuth } from '../providers/AuthProvider';
+
+export const GET_LOGGABLE_EVENTS_FOR_USER = gql`
+    query GetLoggableEventsForUser($userId: String!) {
+        user(userId: $userId) {
+            id
+            loggableEvents {
+                id
+                ...LoggableEventFragment
+            }
+            eventLabels {
+                id
+                ...EventLabelFragment
+            }
+        }
+    }
+    ${EventLabel.fragments.eventLabel}
+    ${LoggableEventCard.fragments.loggableEvent}
+`;
 
 /**
  * LoggableEventsGQL component that fetches and displays loggable events and event labels for the authenticated user.
+ * Fetched data is stored in the Apollo Client cache and child components can access it via fragments.
  * This component controls the loading state and error state of the main view based on the fetch status.
  */
 const LoggableEventsGQL = () => {
@@ -14,7 +35,8 @@ const LoggableEventsGQL = () => {
 
     invariant(user, 'User is not authenticated');
 
-    const { loading, error } = useLoggableEventsForUser(user, {
+    const { loading, error } = useQuery(GET_LOGGABLE_EVENTS_FOR_USER, {
+        variables: { userId: user.id },
         // In offline mode, only read from cache, don't try network
         fetchPolicy: isOfflineMode ? 'cache-only' : 'cache-and-network'
     });
